requirements.txt
psycopg2-binary
beautifulsoup4
requests


